// src/api/content/types/content.type.ts
import { PledgeType } from './pledge.type';
import { ProgrammeType } from './programme.type';
import { QuizAttemptType, QuizType } from './quiz.type';
import { StoryStopType } from './story-stop.type';

export type ContentType = {
  id: string;
  uuid: string;
  priority: number;
  title: string;
  description: string;
  format: string;
  content: string;
  badgeCriteria: string;
  quizDisplayCriteria: string;
  category: string;
  isFeatured: boolean;
  isDeleted: boolean;
  programmeId: string | null;
  programme?: ProgrammeType | null;
  badgeId: string | null;
  badge?: {}; // TODO: Define badge type, then virtual populate
  quizId: string | null;
  quiz?: QuizType | null;
  criticalThinkingQuizId: string | null;
  criticalThinkingQuiz?: QuizType | null;
  thumbnail: string;
  season: number;
  cast: string[];
  quizAttempt?: QuizAttemptType | null;
  isPreviousCompleted?: boolean; // Virtual field to indicate if previous content's quiz is completed

  createdAt?: Date; // auto-generated by MongoDB
  updatedAt?: Date; // auto-generated by MongoDB

  pdfUrl: string | null; // Optional field for PDF URL

  storyStopId: string | null;
  storyStop: StoryStopType | null;
  pledgeId: string | null;
  pledge: PledgeType | null;
};

export enum ContentBadgeCriteriaEnum {
  END = 'end',
  MIDPOINT = 'midpoint',
  START = 'start',
  QUIZ = 'quiz',
}

export enum ContentQuizDisplayCriteriaEnum {
  END = 'end',
  MIDPOINT = 'midpoint',
  START = 'start',
}

export type ViewingHistoryStatType = {
  name: string;
  watched: number;
  totalContent: number;
};